Fixup 32/64 packet lengths.
git-svn-id: file:///tmp/tmp.jyKeACR6iO/svn-version/trunk@86 f84e74b7-95bb-4e84-9b5c-02b1165c2552

+++ afp.h
@@ -432,7 +432,7 @@ int afp_volclose(struct afp_volume * volume);
 
 
 int afp_setforkparms(struct afp_volume *volume,
-        unsigned short forkid, unsigned short bitmap, unsigned int len);
+        unsigned short forkid, unsigned short bitmap, unsigned long len);
 
 int afp_byterangelockext(struct afp_volume * volume,
         unsigned char flag,
